+++ src/player/sound-source/create-note.js
@@ -11,9 +11,9 @@ export default function createNote(option) {
     let note2;
 
     // 音色の設定 //
-    switch (this.channels[note.channel][0]/10 || option.instrument) {
+    switch (this.channels[note.channel][0]*1000 || option.instrument) {
         // Sine
-        case 0.1:
+        case 1000:
         case  6: case 15: case 24: case 26: case 46: case 50: case 51:
         case 52: case 53: case 54: case 82: case 85: case 86:
         {
@@ -22,7 +22,7 @@ export default function createNote(option) {
             break;
         }
         // Square
-        case 0.2:
+        case 2000:
         case  4: case 12: case 13: case 16: case 19: case 20: case 32: case 34: case 45: case 48: case 49:
         case 55: case 56: case 57: case 61: case 62: case 63: case 71: case 72: case 73: case 74: case 75:
         case 76: case 77: case 78: case 79: case 80: case 84:
@@ -32,7 +32,7 @@ export default function createNote(option) {
             break;
         }
         // Sawtooth
-        case 0.3:
+        case 3000:
         case  0: case  1: case  2: case  3: case  6: case  7: case 17: case 18: case 21: case 22: case 23:
         case 27: case 28: case 29: case 30: case 36: case 37: case 38: case 39: case 40: case 41: case 42:
         case 43: case 44: case 47: case 59: case 64: case 65: case 66: case 67: case 68: case 69: case 70:
@@ -42,7 +42,7 @@ export default function createNote(option) {
             break;
         }
         // Triangle
-        case 0.4:
+        case 4000:
         case  8: case  9: case 10: case 11: case 14: case 25: case 31: case 33: case 35: case 58: case 60:
         case 83: case 88: case 89: case 90: case 91: case 92: case 93: case 94: case 95:
         {
